new additions

+++ Interest.java
@@ -8,11 +8,11 @@ public class Interest {
         Scanner vals=new Scanner(System.in);//creates scanner object
 
         System.out.println("please enter the interest rate");
-        int interest=vals.nextInt();
+        float interest=vals.nextFloat();
         System.out.println("please enter the initial amount");
-        int initial_amount=vals.nextInt();
+        float initial_amount=vals.nextFloat();
         System.out.println("please enter the time");
-        int time=vals.nextInt();
+        float time=vals.nextFloat();
 
         float simple_interest=time*initial_amount*interest;
         System.out.println("simple interest for the past "+time+" years with the initial as "+initial_amount+" and initial rate as "+ interest+ " is "+ simple_interest);
